requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
ale-py==0.10.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.1.0
apache-airflow==2.8.4
apache-airflow-providers-common-io==1.3.0
apache-airflow-providers-common-sql==1.11.1
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.10.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-smtp==1.6.1
apache-airflow-providers-sqlite==3.7.1
apispec==6.6.0
appdirs==1.4.4
argcomplete==3.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
audioread==3.0.1
Automat==22.10.0
Babel==2.13.1
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.8.2
cachelib==0.9.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chess==1.10.0
click==8.1.7
clickclick==20.10.2
cloudpickle==3.1.0
cmdstanpy==1.2.0
colorama==0.4.6
colorlog==4.8.0
comm==0.2.0
ConfigUpdater==3.2
connexion==2.14.2
constantly==23.10.4
contourpy==1.2.0
cron-descriptor==1.4.3
croniter==2.0.3
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
docutils==0.20.1
email_validator==2.2.0
executing==2.0.1
Farama-Notifications==0.0.4
fastapi==0.115.6
fastjsonschema==2.19.0
ffmpeg==1.4
ffmpy==0.5.0
filelock==3.13.1
flappy-bird-gymnasium==0.4.0
Flask==3.0.3
Flask-AppBuilder==4.3.11
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.45.1
fqdn==1.5.1
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
git-filter-repo==2.45.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.23.4
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
google-re2==1.1
googleapis-common-protos==1.63.0
gradio==5.9.1
gradio_client==1.5.2
greenlet==3.0.3
grpcio==1.59.3
gunicorn==21.2.0
gymnasium==1.0.0
h11==0.14.0
h5py==3.10.0
holidays==0.41
html5lib==1.1
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.27.0
hyperlink==21.0.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.11.0
importlib-resources==6.1.1
incremental==22.10.0
inflection==0.5.1
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.15.0
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
limits==3.10.1
linkify-it-py==2.0.3
llvmlite==0.41.1
lockfile==0.12.2
lxml==5.1.0
Mako==1.3.2
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
multitasking==0.0.11
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
ninja==1.11.1.1
nltk==3.9.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.9.0.80
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.10.12
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
pathspec==0.12.1
peewee==3.17.0
pendulum==3.0.0
pgn-downloader==0.2.2
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.4.0
pooch==1.8.0
prison==0.2.1
proglog==0.1.10
prometheus-client==0.19.0
prompt-toolkit==3.0.41
prophet==1.1.5
Protego==0.3.0
proto-plus==1.25.0
protobuf==4.23.4
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.10.4
pydantic_core==2.27.2
PyDispatcher==2.0.7
pydub==0.25.1
pygame==2.6.0
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.20
python-nvd3==0.15.0
python-slugify==8.0.4
pytube==15.0.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
queuelib==1.6.2
referencing==0.31.1
regex==2024.11.6
requests==2.32.3
requests-file==2.0.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rich-argparse==1.4.0
rpds-py==0.13.2
rsa==4.9
ruff==0.8.4
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.11.4
Scrapy==2.11.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
service-identity==24.1.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.4.4
stable_baselines3==2.4.0
stack-data==0.6.3
stanio==0.3.0
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io==0.29.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.3.0
terminado==0.18.0
text-unidecode==1.3
threadpoolctl==3.2.0
time-machine==2.14.1
tinycss2==1.2.1
tldextract==5.1.1
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.47.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typer==0.15.1
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2023.3
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal-pathlib==0.1.4
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.34.0
virtualenv==20.25.1
w3lib==2.1.2
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==14.1
Werkzeug==3.0.3
widgetsnbextension==4.0.9
wrapt==1.14.1
WTForms==3.1.2
yarl==1.9.4
yfinance==0.2.36
youtube-dl==2021.12.17
zipp==3.18.1
zope.interface==6.2
